+++ ui/opensnitch/service.py
@@ -116,7 +116,8 @@ class UIService(ui_pb2_grpc.UIServicer, QtWidgets.QGraphicsObject):
             self._cfg.getBool(self._cfg.DEFAULT_IGNORE_RULES),
             self._cfg.getInt(self._cfg.DEFAULT_IGNORE_TEMPORARY_RULES)
         )
-        self._nodes.delete_rule_by_field(Config.DURATION_FIELD, Config.RULES_DURATION_FILTER)
+        if self._cfg.getBool(self._cfg.DEFAULT_IGNORE_RULES):
+            self._nodes.delete_rule_by_field(Config.DURATION_FIELD, Config.RULES_DURATION_FILTER)
 
     # https://gist.github.com/pklaus/289646
     def _setup_interfaces(self):
